refactor: move incognito mode settings to more settings dialog
